cloudfunctions/imagemagick/requirements.txt
google-cloud-storage==1.13.0
Wand==0.4.5
